requirements.txt
gdown==4.7.1
PyYAML==6.0.2
opencv-python==4.11.0.86
matplotlib==3.10.1
scikit-learn==1.6.1
supervision==0.25.1
torch==2.1.2+cu118
torchvision==0.16.2+cu118
torchaudio==2.1.2+cu118
--extra-index-url https://download.pytorch.org/whl/cu118
pynvml==11.5.0
rfdetr==1.1.0
numpy==1.24.4
tensorboard==2.14.0
